chore: move container structures
